:truck: Move Api functionality to dedicated folder
