+++ platforms/platform-alexa/src/api/index.ts
@@ -1,0 +1,2 @@
+export * from './AlexaApi';
+export * from './CustomerProfileApi';
